requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
loguru==0.6.0
mangust228==0.6
multidict==6.0.4
numpy==1.24.2
pandas==1.5.3
playwright==1.32.1
pydantic==1.10.7
pyee==9.0.4
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
telebot==0.0.5
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
